t0000-basic: Add ls-tree recursive test back.
When we updated ls-tree recursive output to omit the tree nodes,
30c9663581218afaa5b073bd0bdbe86a09b2f4d8 adjusted the old test
so that we do not expect to see trees in its output.  Later,
with 23599da763bfa7b97318b7cd548a9cd8e6fcc3c9, we added back the
ability to show both with -t option, but we forgot to update the
test as well.

Signed-off-by: Junio C Hamano <junkio@cox.net>