gitk: Move "pickaxe" find function to highlight facility
This removes the "Files" and "Pickaxe" parts of the "Find" function,
so Find is now just about searching the commit data.  We now highlight
the commits that match the Find string (without having to press Find),
and have a drop-down menu for selecting whether the git-diff-tree based
highlighting is done on paths or on adding/removing a given string.

Signed-off-by: Paul Mackerras <paulus@samba.org>